CryptGenRandom expects DWORD

+++ common/randombytes.c
@@ -100,7 +100,7 @@ static int randombytes_win32_randombytes(void *buf, const size_t n) {
         return -1;
     }
 
-    tmp = CryptGenRandom(ctx, n, (BYTE *) buf);
+    tmp = CryptGenRandom(ctx, (unsigned long)n, (BYTE *) buf);
     if (tmp == FALSE) {
         return -1;
     }
